refractor: remove selectedLayers loop
The selected layers were never used, so there is no reason to loop
through them.

CompItem.selectedProperties contains the all selected properties
for all selected layers, and it is not possible to selected a property
without selecting the layer.

Having the selectedLayers loop would also cause the script to loop
through all selected properties in the comp as many times as the there
were selected layers. Had it not been for the i variable being used
multiple times inside the loop, for looping over the keys.